New Features, Improvements, and Changes
New Features:
- Added Bluetooth LE as a DeviceInterface to connect to a Dexcom
Receiver.

Changes:
- The Transmitter ID isn't queried until the caller specifically
requests it rather than upon connection.
- GetManufacturingData() doesn't attempt to query device properties
until those properties are requested. This speeds up the connection
process over both USB and Bluetooth.
- The TransmitterID property has been replaced by the GetTransmitterID
function to better reflect the fact that it needs to be queried
seperately.
- Receiver.ConnectToDevice() has been replaced by Receiver.Connect(). An
overload of this function without parameters was added to attempt to
reestablish a previous connection.
- Querying databases with an earliest record time previously returned
data that was up to an hour earlier than that time. This was to ensure
data points weren't missed. This has been shortened to 15 minutes.